requirements.txt
langchain
langchain-openai
langchain-community
openai
streamlit
python-dotenv
